argos-adf/requirements.txt
jsonschema==3.0.2
matplotlib==3.1.1
numpy==1.17.1
requests==2.22.0
slackclient==2.1.0